// src/app/features/colleges/Al-Alsun/model/journal.model.ts
export interface JournalInfo {
//   title: string;
  issuedBy: string;
  establishmentDate: string;
  languages: string;
  frequency: string;
  type: string;
  issuesPublished: string;
  description: string;
  coverImage: string;
}

export interface JournalButton {
  label: string;
  link: string;
  isPrimary: boolean;
}